third_party/WebKit/LayoutTests/external/wpt/webstorage/event_session_url_usvstring.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>WebStorage Test: sessionStorage event - url USVString</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>event_session_URL USVString test</h1>
    <div id="log"></div>
    <script>
        async_test(function(t) {
            function onStorageEvent(event) {
                assert_equals(event.url, window.location.href + "\ufffd");
                t.done();
            }

            window.addEventListener('storage', t.step_func(onStorageEvent), false);

            var event = new StorageEvent('storage', {
                url: window.location.href + '\uD999',
            });

            window.dispatchEvent(event);
        }, "url property test of session event - unpaired surrogate codepoint should be replaced with U+FFFD");
    </script>
 </body>
</html>
